mlx4_en integration: Added Shahar's fix to the steering mode issue
The new ported mlx4_core driver is in B0 steering mode by default and
the old mlx4_en driver is in A0 steering mode by default.
If high_rate_steer == TRUE it means that A0 steering mode is on.
The integration fix is to hard code high_rate_steer to TRUE.

Change-Id: I4265f4266eab9707a13d5c3f83cad83d1bbb3eb6
Signed-off-by: Oded Shanoon <odeds@mellanox.com>