Fixed kernel crash on reload of mlx4_en module
The order of releasing resources in mlxen was wrong, which caused panic on reload of the module.
conf_ctx list should be released before stat_ctx list, otherwise the leafs in conf_ctx list won't be released because of the dependancy.
The fix is to change the order of the releases.

Issue: 227883

Change-Id: I2e26bbb4d0f72094eb53027bdead64ab753e29bc
Signed-off-by: Daria Zasipko <dariaz@mellanox.com>