Fixed kernel crash on reload of mlx4_ib and IPoIB modules
class_attributes were not released as they should have been, which caused some leaves not to be released.
In case that one of this modules was reloaded, the command sysctl -a caused a crash.
Added release of class_attributes in function ib_device_unregister_sysfs in sysfs.c and in mlx4_ib_remove in main.c (in each function different leaves are removed).

Change-Id: I8a4a6b84c2d573e7b1333ac15a2fc8fe4fc2bb05
Signed-off-by: Daria Zasipko <dariaz@mellanox.com>